# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nimiq Mini Apps API",
    "version": "1.0.0",
    "description": "REST API for the [Nimiq Mini Apps](https://nimiqminiapps.com) community directory.\n\n**Submitting an app** — `POST /api/apps/submit` (wallet login required). Rate limit: **5 requests per hour per IP**.\nRequired field `submitter_contact` (Telegram, email, etc.) is stored privately and omitted from public responses.\n\n**Open in Nimiq Pay** — listed apps expose `open_url` as `https://nimpay.app/miniapps/open/{domain}` (domain without `https://`).\n\nMachine-readable spec: `GET /openapi.json` or `GET /openapi.yaml`.\n",
    "contact": {
      "name": "Nimiq Mini Apps",
      "url": "https://nimiqminiapps.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/NimMiniApps/NimiqMiniApps/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://api.nimiqminiapps.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Search users by display name or wallet address (for the developer picker)",
        "operationId": "adminSearchUsers",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Prefix match on display_name or wallet_address; empty returns []",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "wallet_address": {
                        "type": "string"
                      },
                      "display_name": {
                        "type": "string",
                        "nullable": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/apps/{slug}/owners": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Add a co-owner wallet (admin, no membership requirement)",
        "operationId": "adminAddAppOwner",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "wallet_address"
                ],
                "properties": {
                  "wallet_address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner added (or already present)"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/apps/{slug}/owners/{wallet}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Remove a co-owner wallet (admin — can remove the last owner to fully unclaim)",
        "operationId": "adminRemoveAppOwner",
        "security": [
          {
            "adminBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owner removed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set ADMIN_TOKEN on the server; send as Authorization Bearer token"
      },
      "walletCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wallet_session",
        "description": "Signed wallet-session cookie set by POST /api/auth/verify"
      }
    },
    "parameters": {
      "slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Invalid or missing admin token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}